fix typo in doc
Interestingly "r1 and r2" does return a Region, namely r2.
This is due to the short-circuiting behavior in python:
	https://docs.python.org/2/library/stdtypes.html#boolean-operations-and-or-not
which returns r2 when bool(r1) is True, i.e., when r1 is a Region with non-zero volume.
On the contrary, r1 and not r2 is Boolean if bool(r1) is True, because it returns not r2, which is Boolean valued.