rename methods: [Convex]Polytope.reduce -> .reduction
to emphasize that a new object is returned (namely the reduction of that polytope), whereas the object to which the method is bound remains unaffected.
In contrast, the verb "reduce" would suggest actively operating and changing the polytope, which isn't the case.
This also pertains to (suggested) immutability of polytopes.