rm "show" arg from plot_partition
because it is not so explicit and can hide plots when not expected.
Also, it leads to propagation of too many args in functions that use it.

Use instead a ternary one-liner in the top-level user script to throttle
plotting:

show = True
...
plot_partition(...) if show else 0

This approach is more explicit and readable and it doesn't consume
significantly more space.