API: rename attribute of class `Partition` to `domain`
from `Partition.set`. This renaming avoids using the same
identifier as the builtin function `set`. It also avoids
confusion that can result from interpreting the attribute's
name to mean that it is of type `set` in Python.

This attribute does not appear to be used anywhere in the
packages `tulip` and `polytope`, except within the module
`polytope.prop2partition`.