MAINT: Rename functions to tranlation, rotation, _translate, and _rotate.
Apparently convention is that object methods return a copy of the object and
that methods are described using the noun form when doing so and the verb
form when modifying in place. These functions have been renamed and altered
to return copies to reflect this convention. The helper functions have been
hidden from the public API using underscore.