API: return `None` from `plot` methods if polytope has zero volume
Logging an error suggests that this case is abnormal.
So, returning an axes handle doesn't follow usual conventions.

On the other hand, it is better to not raise an exception,
because sometimes a flat polytope can occur, and it may be
useful to still plot in these cases, e.g., during debugging.