BUG: `np.vstack` can't stack 0 with 1D array
If `len(mi) > 2`, then `np.vstack` fails because the sizes of `[0]`
and `mi` do not match. This is fixed by using `np.concatenate`
instead because it can join the two arrays along the 0th dimension.

I believe this maintains the intent of this line of code which is to
append `0` to the front of the first `len-1` elements of `mi`. It
doesn't matter that the resulting array is 1D instead of a vertical
2D array because `np.cumsum` without parameters operates on the
flattened version of its input and returns a 1D array.

However, I do not understand any larger impact this change may have
on `region_diff` because the documentation does not explain how the
algorithm is intended to work.