CI: do not use Miniconda, `pip install --only-binary=scipy,numpy`
The version of Miniconda used on Travis CI contained Python 2.7,
as a result there was incorrect behavior on Python 3.x environments.
One effect was installation of wheels whose major Python version
didn't match the environment. Introduced in [1].

A relevant correction is [2], though that version of Miniconda
contains a specific Python 3 release (3.6.2 as of this commit),
leading to mismatch of the minor Python version for other
Python 3.x environments on Travis CI.

By installing wheels from PyPI for `scipy` and `numpy` [3] the build
takes a short time, as opposed to building them from source,
which originally motivated using Miniconda.

[1] 370c7b5571722352f19548ef083aecff7f5a09d5
[2] https://github.com/tulip-control/tulip-control/commit/60d0a4b54cdd8a7f26000e9691672a9e14745ab5
[3] https://github.com/travis-ci/travis-ci/issues/2650#issuecomment-252285413