BUG: return 1-D `numpy.array` from `lpsolve` when using GLPK
`numpy.squeeze(x)` removes all dimensions that have size 1.
So an array with only one dimension of size 1 becomes
a zero dimensional array, which does not match the API
of `scipy.optimize.linprog`.

`cvxopt` returns a 2D `cvxopt.matrix`. If `cvxopt` returned a 2D
`numpy.array`, then `numpy.squeeze(x, axis=1)`
would squeeze the second dimension (which is always of size 1),
and yield the desired result.

However, applying `numpy.squeeze` to a 1x1 `cvxopt.matrix`,
even with `axis=1`, produces a zero dimensional array.
Introduced in: 79d85ec95a32819534011756bb0200c824ae2148

This motivates the change in this commit.
A number of other approaches are possible, as discussed below.
The choice of change was based on profiling.

close #47

Profiling of alternatives
=========================

A speedup within `lpsolve` is desirable, due to the frequency with
which `lpsolve` is called. Given that abstraction computations involve
many small optimization problems, it may be the case that delays within
`lpsolve` due to Python's behavior are comparable with the time it
takes for GLPK to solve a small problem.

Comparison to list comprehension
--------------------------------

```python
import cvxopt
import numpy as np

a = cvxopt.matrix([1.0, 2.0, 3.0, 4.0, 5.0])

>>> %%timeit
>>> c = np.array([u for u in a])
2.94 µs ± 5.81 ns per loop (mean ± std. dev. of 7 runs, 100000 loops
each)

>>> %%timeit
>>> c = np.fromiter(a, dtype=np.double)
2.1 µs ± 32.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops
each)
```

So there is roughly a 37% speedup compared to list comprehension.

Comparison to matrix-based approach
-----------------------------------

A `matrix`-based approach yields a `matrix` instead of a 1D array, so
it differs from the API of `scipy.optimize.linprog` (the other branch).
For completeness of the comparison:

```python
>>> %%timeit
>>> c = np.matrix(a).transpose()
13 µs ± 275 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

So this approach takes more time ((13 µs - 275 ns) / (2.1 µs + 32.2 ns)
= 5.96).

Comparison to conversion then squeezing
---------------------------------------

Another approach is to first convert from `cvxopt.matrix` to
numpy.array`, and then squeeze:

```python
>>> %%timeit
>>> c = np.array(a).squeeze(axis=1)
3.51 µs ± 249 ns per loop (mean ± std. dev. of 7 runs, 100000 loops
each)
```

So this approach is 52% slower than `np.fromiter`.

About the assertion
===================

According to the documentation we don't need the assertion.
However, the documentation of `scipy.optimize.linprog` is not explicit
about types (though the only dense real type in `cvxopt` is double).
The purpose of the assertion is to catch type errors early,
e.g., if the API of `scipy` or `cvxopt` changes in the future.