API: add arg `solver` to function `lpsolve`
Explicit is better than implicit [PEP 20].
Selecting a solver only implicitly via module state is undesired.

In presence of `cvxopt.glpk`, any user that wants to use
`scipy.optimize.linprog` is forced to modify module state
(the variable `polytope.polytope.lp_solver`). The purpose
of this variable is to record what is available in the
runtime environment, for the module's own decisions.
The variable should not become global state that the user
needs to change in order to achieve desired behavior.

This change allows the user to select the desired solver,
and uses knowledge of the runtime environment only to
select a reasonable default solver from what is installed.

`scipy.optimize` is imported in all cases, in order to
allow the user to select the desired solver.
`scipy` is required in `install_requires`, thus this
import should succeed.

close #17