MAI: replace `except:` with `except Exception:`
as recommended by PEP8:

"When catching exceptions, mention specific exceptions
whenever possible instead of using a bare except: clause.
...
A bare except: clause will catch SystemExit and
KeyboardInterrupt exceptions, making it harder to
interrupt a program with Control-C, and can disguise
other problems. If you want to catch all exceptions
that signal program errors, use except Exception:
(bare except is equivalent to except BaseException:)."

The statements `except:` and `except Exception` are
inequivalent. Nevertheless, the changes are appropriate
in the cases addressed by this commit.

This change was motivated by a suggestion of `autopep8 -a`
to use `except BaseException:` instead of `except:`,
which are equivalent.