API: Remove `contains` and move `are_inside` to module level
In order to keep backwards compatability and keep the API
simplier, `contains` was removed because it was a redundant
method to `are_inside`. However, `are_inside` was only implemented
as a method of the `Polytope` class. Thus `are_inside` was moved
to the module level so it can be called like it's relative,
`is_inside`.

Additional documentation was added for `are_inside`, `is_inside`,
and `__contains__` to make finding these related methods easier.