API: rename variable `lp_solver` to `default_solver`
This change has two goals:

1. readability

2. change of meaning: the variable `lp_solver` was used
   to record both the default choice of solver, and the
   best installed solver. This was wrong, even for 2 solvers
   in total.

   The error is due to the following case: With GLPK
   installed, setting `lp_solver` to SciPy does change
   the default solver, but also causes direct calls
   to `lpsolve` that request GLPK to fail, because the
   module now thinks that GLPK isn't installed.

   A second variable is needed in addition to
   `default_solver`, in order to record the installed
   solvers. Using two variables is necessary also for
   supporting more than two solvers.

So the change modifies the API both due to renaming a variable
used for configuration, and because *what* this variable
configures changes.

Notice also that the check `solvers.default_solver != 'glpk'`
is more restrictive than necessary, and will be replaced by
a check that GLPK is installed.