API: record env in the variable `installed_solvers`
This variable addresses the issue of checking whether
a requested solver is available, independently of what
solver is selected as default. The default solver is
stored in the variable `default_solver`.

Any additional solvers supported in the future will
be added to the set `installed_solvers` if successfully
imported.

Choice of default solver
========================

The presence of those solvers need not affect the
variable `default_solver`, unless the user explicitly
selects one of those solvers as default (e.g., MOSEK).

In contrast, the presence of GLPK does affect the
default solver. The choice of default solver was
shifted outside of the `try...except` in order to
clearly separate the two operations of:

1. recording the runtime environment
2. choosing a default solver from those present

Choice of exception type
========================

Unavailability of a requested solver is announced with
a `RuntimeError`. Previously, an `ImportError` was
raised.

The change of error type is motivated by the documentation
of `ImportError`, which is specific to the `import`
statement, which is underlines by the `name` and `path`
attributes [1]:

> When set they represent the name of the module that
> was attempted to be imported and the path to any file
> which triggered the exception, respectively.

Absence of a solver may be due to reasons other than
a single module failing to import. For example,
when MOSEK will be supported, it requires that both
`cvxopt` and MOSEK's Python bindings be installed.

[1] https://docs.python.org/3/library/exceptions.html#ImportError

Misc
====

The assertion that `scipy` is available within the
function `_solve_lp_using_scipy` can be omitted,
but was added for uniformity and as a local guard.