TST: update to `numpy >= 1.14.0`
The formatting of numpy arrays as strings has changed in
`numpy >= 1.14.0`. Even though this affects a test of `polytope`,
the difference concerns only whitespace. So there appears to be
no reason yet for bumping the required `numpy` version.

So this test will fail in environments with `numpy < 1.14.0`,
even though `polytope` works with earlier `numpy` versions too.