Checking for linearly dependent rows a bit smarter
Hyperplanes that represent the same constraint have equal parameters
when they are normed. The constant in the b vector can be included when
the norm is calculated. However, when the comparison of two hyperplanes
are made with some absolute tolerance it is better to leave b out of the
norm.

Also, instead of removing an arbitrary hyperplane, try to keep the most
constraining one.