REL: rm support for Python 2.7
The CI tests for Python 2.7 for commit
50a5ee75e52f40dd29a651f89e99690ec548aba2
failed, because the function `numpy.random.default_rng`
does not exist on the version of `numpy` that was installed
on Python 2.7.

The function `numpy.random.default_rng` was [added](
    https://github.com/numpy/numpy/commit/6658517495076c25f581c7093c39ad444f458b6f)
to `numpy` in version 1.17.0.

The reason why the tests failed was that the last `numpy` version
compatible with Python 2.7 was released on Dec 29, 2019
[`numpy == 1.16.6`](https://pypi.org/project/numpy/1.16.6/).
Larger versions of `numpy` do not support Python 2.7 since Jul 26, 2019
([`numpy == 1.17.0`](https://pypi.org/project/numpy/1.17.0/)).