API: require that `res` contain values >= 1
in the function `polytope.polytope.grid_region`.
At least one value is required, in order to
ensure that the (unfiltered) grid over the
bounding box is nonempty
("unfiltered" here means the grid before keeping
only those points that are inside the given polytope).

Otherwise, the function `numpy.meshgrid` would
return an empty grid, which would most likely
be unintended. In other words, there is no reason
for calling the function `grid_region` with
a number of 0 points for the unfiltered grid.

In case such a zero arises in user code, and means
that the grid would have been empty, then this
should be checked in user code, and the function
`grid_region` not called at all in this case
(explicit is better than implicit, PEP 20).

Also:
- mention this requirement in the function's docstring
- test that the exception `ValueError` is raised when expected.

This change formalizes what was already the intended
purpose of the function `grid_region`, so this change
is not mentioned in the changelog.