MAI: call `networkx.to_numpy_array`, instead of `networkx.to_numpy_matrix`
because the function `networkx.to_numpy_matrix` has been deprecated
in `networkx >= 2.6` (under development) [commit][1]. As recommended
by the deprecation message of `networkx`, use instead the function
[`networkx.to_numpy_array`][2]. The keyword parameter `nodelist`
has the same description in both of the functions `to_numpy_array`
and [`to_numpy_matrix`][3].

Moreover, there is no need to convert the array to a matrix,
because on the next line, the matrix was converted to an array.
Remove that unnecessary conversion.

[1]: https://github.com/networkx/networkx/commit/ddf707dc0c35edf5bda884de8534899af945ad78
[2]: https://networkx.org/documentation/latest/reference/generated/networkx.convert_matrix.to_numpy_array.html
[3]: https://networkx.org/documentation/stable/reference/generated/networkx.convert_matrix.to_numpy_matrix.html